+++ src/main/java/roomescape/repository/JdbcThemeRepository.java
@@ -10,6 +10,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
+import roomescape.domain.ThemeRepository;
 
 @Repository
 public class JdbcThemeRepository implements ThemeRepository {
