+++ src/main/java/roomescape/service/ThemeService.java
@@ -9,8 +9,8 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.Theme;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ThemeRepository;
+import roomescape.domain.ReservationRepository;
+import roomescape.domain.ThemeRepository;
 
 @Service
 @Transactional(readOnly = true)
