+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -22,8 +22,8 @@ import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.dto.ReservationTimeResponse;
 import roomescape.dto.ThemeResponse;
-import roomescape.repository.ReservationTimeRepository;
-import roomescape.repository.ThemeRepository;
+import roomescape.domain.ReservationTimeRepository;
+import roomescape.domain.ThemeRepository;
 
 class ReservationControllerTest extends BaseControllerTest {
